backend/pyproject.toml
[tool.poetry]
name = "shortener"
version = "0.1.0"
description = ""
authors = ["Roman Vlasenko <klavionik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.2.1"
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.23"
isort = "^5.12.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
